add helper function to mock commander
